python/flask-demo-2/pyproject.toml
[tool.poetry]
name = "flask-demo-2"
version = "0.1.0"
description = ""
authors = ["Hardik Choksi <hardik.choksi@middleware.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
langchain-groq = "0.0.1"
langchain-openai = "0.0.6"
middleware-apm = {path = "/home/hardik/Projects/agent-apm-python", develop = true}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
